sr-iov: Replace whitelisted PFs with VFs
The Sunbeam interactive menu will provide a list of SR-IOV capable
PFs, letting the user pick which ones to expose to Openstack
tenants.

Nova exposes the associated VFs implicitly only if the PF is
specified by name (deprecated) or by PCI address, but not when
the product/vendor IDs are provided.

Furthermore, Nova no longer allows the PF *and* its VFs to be
whitelisted.

https://github.com/openstack/nova/blob/2010536d12b684a425ff00068da4317b4efd4951/doc/source/admin/pci-passthrough.rst?plain=1#L58-L61

As such, we'll process the PCI whitelist and replace PFs with the
corresponding VFs, if any.