list-nics: include attached PFs and VFs
Right now, "list-nics" won't include PFs and VFs that have already
been attached to Openstack instances.

This affects the Sunbeam SR-IOV configuration dialog as well as
the openstack-hypervisor configuration hook, which may not
whitelist all the required VFs.

To address this, we're going to iterate over the PCI device list,
looking for omitted network adapters.