Rewrite devspec module
Sunbeam needs to check if a given PCI device is whitelisted.

The imlementation needs to be compatible with Nova, however
we can't import Nova code since it's a large dependency and we
can't copy Nova modules either since that could break the Canonical
license agreement. Both projects are currently Apache 2 licensed,
however the Canonical CLA demands re-licensing rights.

That being considered, we'll rewrite the devspec module, greatly
simplifying it.

Nova no longer allows the pci device specs to include device names,
we can drop that.

We'll also avoid explicitly checking if a device is a PF and rely
solely on addresses. The advantage is that this can be used
remotely, without querying any local devices.