Handle dpdk physical ports
This change adds two new settings:

* `network.ovs-dpdk-ports` list of interface names that will be
  used with DPDK. The interfaces will be bound to the DPDK-compatible
  driver (vfio-pci by default) and corresponding DPDK ports will be
  created in OVS. Any bridges and bonds that use those interfaces
  will be updated with the resulting DPDK ports, based on the
  MAAS/Netplan configuration.
* `network.dpdk-driver` The DPDK driver used to control physical network interfaces.
  Possible values: `vfio-pci` (default) or `uio-pci-generic`.

The dpdk port configuration extracted from Netplan will be stored
as a snap setting: `internal.dpdk-port-mappings`. This allows
the snap to recreate the dpdk ports after the original
interfaces are removed from the Netplan configuration.

At the same time, the charm can use `internal.dpdk-port-mappings`
to obtain the PCI addresses of the dpdk physical ports and determine
which NUMA nodes to use when allocating OVS cores and huge pages.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>