sr-iov: avoid duplicate VF specs
At the moment, nova.conf contains duplicate PCI device specs for
whitelisted SR-IOV VF.

The reason is that for each whitelisted PF,
"process_whitelisted_sriov_pfs" adds the VFs to the whitelist and
the PF to the exclusion list.

"apply_exclusion_list" then processes the PF spec, sees that the
device is excluded but also detects the matching VFs, which
are added again to the whitelist.

We'll avoid this by modifying "apply_exclusion_list" to ignore
parent PFs since the list of VFs is populated separately.